pyproject.toml
# PEP 621 project metadata
[project]
name = "mochirag"
version = "0.1.0"
description = "MochiRAG: A multi-tenant RAG application"
authors = [
    {name = "Jules", email = "jules@example.com"},
]
readme = "README.md"
requires-python = ">=3.10, <4.0"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    "python-dotenv>=1.0.1",
    "pydantic[email]>=2.8.2",
    "pydantic-settings>=2.3.4",
    "sqlalchemy>=2.0.31",
    "passlib[bcrypt]>=1.7.4",
    "python-jose[cryptography]>=3.3.0",
    "langchain-core>=0.2.10",
    "pyyaml>=6.0.1",
    "python-multipart>=0.0.9",
    "langchain-community>=0.2.5",
    "chromadb>=0.5.3",
    "unstructured[md]>=0.14.8",
    "pypdf>=4.2.0",
    "sentence-transformers>=3.0.1",
    "langchain-ollama>=0.2.1",
    "langchain-huggingface>=0.0.3",
    "streamlit>=1.36.0",
    "httpx>=0.27.0",
    "langchain-chroma>=0.1.2"
]

[project.optional-dependencies]
test = [
    "pytest>=8.2.2",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "anyio[trio]>=4.0.0",
]
dev = [
    "ruff>=0.5.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
py-modules = [] # py-modules is not used
packages = ["backend", "core", "config"]

[tool.poetry]
package-mode = false

[tool.poetry.scripts]
start = "uvicorn backend.main:app --reload --port 8000"
test = "pytest"

[tool.pytest.ini_options]
pythonpath = [
  "."
]
